+++ src/app/js/ont/rj-viewer.js
@@ -17,13 +17,14 @@
       controller: RjDataViewerController,
       scope: {
         uri:  '=',
-        rj:   '='
+        rj:   '=',
+        triples: '='
       }
     }
   }
 
-  RjDataViewerController.$inject = ['$scope', 'uiGridConstants'];
-  function RjDataViewerController($scope, uiGridConstants) {
+  RjDataViewerController.$inject = ['$scope'];
+  function RjDataViewerController($scope) {
     debug = debug || $scope.debug;
     $scope.debug = debug;
     if (debug) console.debug("++RjDataViewerController++ $scope=", $scope);
@@ -45,69 +46,12 @@
     });
     triples = _.sortBy(triples, "subjectUri");
 
-    // TODO proper filtering; for now all links external
-    var mklinksOnlyExternal =
-      '<div class="ui-grid-cell-contents">' +
-      '<span ng-bind-html="row.entity[col.field] | mklinksOnlyExternal"></span>'
-      + '</div>';
-
-    // http://ui-grid.info/docs/#/tutorial/209_grouping
-    var groupingCellTemplate =
-      '<div><div ng-if="!col.grouping ' +
-      '|| col.grouping.groupPriority === undefined ' +
-      '|| col.grouping.groupPriority === null ' +
-      '|| ( row.groupHeader && col.grouping.groupPriority === row.treeLevel )"' +
-      ' class="ui-grid-cell-contents"' +
-      ' title="TOOLTIP">{{COL_FIELD CUSTOM_FILTERS}}' +
-      '</div></div>';
-
-    var gridApi;
-
-    $scope.allowGrouping = false;
-    $scope.$watch("allowGrouping", function(allowGrouping) {
-      _.each($scope.gridOptions.columnDefs, function(cd) {
-        cd.cellTemplate = allowGrouping ? groupingCellTemplate : mklinksOnlyExternal;
-      });
-      if (allowGrouping && gridApi && gridApi.treeBase) {
-        gridApi.treeBase.toggleRowTreeState(gridApi.grid.renderContainers.body.visibleRowCache[0]);
-      }
-    });
-
-    $scope.gridOptions = {
-      data: [],
-      columnDefs: [
-        {
-          field: 'subjectUri',
-          //width: '**',
-          displayName: 'Subject',
-          grouping: { groupPriority: 0 },
-          cellTemplate: mklinksOnlyExternal
-        },
-        {
-          field: 'propUri',
-          //width: '**',
-          displayName: 'Predicate',
-          cellTemplate: mklinksOnlyExternal
-        },
-        {
-          field: 'value',
-          //width: '*****',
-          displayName: 'Object',
-          cellTemplate: mklinksOnlyExternal
-        }
-      ]
-      ,enableGridMenu: true
-      ,showGridFooter: true
-      ,enableFiltering: true
-      ,onRegisterApi: function(api) {
-        gridApi = api;
-      }
-    };
-
-    appUtil.updateModelArray($scope.gridOptions.data, triples,
+    $scope.triples = [];
+    appUtil.updateModelArray($scope.triples, triples,
       function(done) {
         if (done) {
           $scope.$parent.$digest();
+          if (debug) console.debug("Done update model array: $scope.triples=", $scope.triples.length);
         }
         else {
           $scope.$digest();